+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/container/MuxerContainer.kt
@@ -4,6 +4,7 @@ import android.media.MediaCodec
 import android.media.MediaFormat
 import android.media.MediaMuxer
 import java.nio.ByteBuffer
+import java.util.concurrent.atomic.AtomicBoolean
 
 /**
  * Wrapper around [MediaMuxer].
@@ -11,37 +12,50 @@ import java.nio.ByteBuffer
  * @param path            Output file path.
  * @param containerFormat A valid [MediaMuxer.OutputFormat] value for the output container format.
  */
-class MuxerContainer(path: String, containerFormat: Int) : IContainerWriter {
-    private val mMuxer = MediaMuxer(path, containerFormat)
-    private var mIsStarted = false
+class MuxerContainer(val path: String, private val containerFormat: Int) : IContainerWriter {
+    private var mMuxer: MediaMuxer? = null
+    private var mStarted = AtomicBoolean(false)
+    private var mStopped = AtomicBoolean(false)
 
     override fun start() {
-        if (mIsStarted) return
+        if (mStarted.get() || mStopped.get()) return
 
-        mMuxer.start()
-        mIsStarted = true
+        mStarted.set(true)
+
+        mMuxer?.start()
     }
 
     override fun stop() {
-        if (!mIsStarted) return
+        if (!mStarted.get() || mStopped.get()) return
 
-        mMuxer.stop()
-        mIsStarted = false
+        mStarted.set(false)
+        mStopped.set(true)
+
+        mMuxer?.stop()
     }
 
-    override fun release() {
-        mMuxer.release()
-        mIsStarted = false
-    }
+    override fun addTrack(mediaFormat: MediaFormat): Int {
+        if (mStarted.get() || mStopped.get()) return -1
 
-    override fun addTrack(mediaFormat: MediaFormat): Int = mMuxer.addTrack(mediaFormat)
+        mMuxer = MediaMuxer(path, containerFormat)
+
+        return mMuxer!!.addTrack(mediaFormat)
+    }
 
     override fun writeSampleData(
         trackIndex: Int,
         byteBuffer: ByteBuffer,
         bufferInfo: MediaCodec.BufferInfo
     ) {
-        if (!mIsStarted) return
-        mMuxer.writeSampleData(trackIndex, byteBuffer, bufferInfo)
+        if (!mStarted.get() || mStopped.get()) return
+
+        mMuxer?.writeSampleData(trackIndex, byteBuffer, bufferInfo)
+    }
+
+    override fun release() {
+        stop()
+
+        mMuxer?.release()
+        mMuxer = null
     }
 }