+++ record_windows/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.5
+* fix: Send messages on main thread (workaround).
+* fix: Automatically remove file when no audio has been written.
+
+## 1.0.4
+* fix: Process path as wide string (UTF-16).
+
 ## 1.0.3
 * fix: Crash when starting the recording when the writer can't be created.